requirements.txt
argh==0.26.2
backports.shutil-get-terminal-size==1.0.0
BeautifulSoup4  # BeautifulSoup==3.2.1
bpython==0.12
certifi==2019.6.16
cssselect==0.8
decorator==4.4.0
dj-database-url==0.4.1
Django  # Django==1.11.25
django-annoying==0.7.6
django-appconf==1.0.3
django-codemirror-widget==0.4.0
django-colorfield==0.1.15
django-compressor==2.3
django-debug-toolbar==1.11
django-debug-toolbar-template-timings==0.6.4
django-easy-select2==1.3.4
django-el-pagination==3.2.4
django-guardian==1.4.9
django-image-cropping==1.2.0
django-js-asset==1.2.2
django-mobile==0.5.0
django-mptt-0.10.0  # django-mptt==0.9.1
django-ranged-response==0.2.0
django-redis-cache==1.6.4
django-simple-captcha==0.5.11
django-static-sitemaps==4.6.0
django-statsd-mozilla==0.3.16
django-supervisor==0.4.0
django-tastypie==0.14.2
django-treebeard==1.61
django-ulogin==0.3
django-widget-tweaks==1.3
django-wysiwyg-redactor==0.5.1
drest==0.9.12
easy-thumbnails==2.5
enum34==1.1.6
envvars==0.3.0
eventlet==0.17.4
feedparser==5.2.1
# Flask==0.9
greenlet  # greenlet==0.4.7
httplib2==0.7.7
iptools==0.6.1
ipython==5.8.0
ipython-genutils==0.2.0
Jinja2==2.6
lxml  # lxml==3.2.1
MarkupSafe==0.18
meld3==2.0.0
mimeparse==0.1.3
mock==1.0.1
MonthDelta==0.9.1
mysqlclient  # MySQL-python==1.2.5
oauthlib==3.0.1
pathlib2==2.3.4
pathtools==0.1.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.0
pipenv==2018.11.26
prompt-toolkit==1.0.16
psycopg2-2.8.3  # psycopg2==2.7.7
ptyprocess==0.6.0
Pygments==1.5
python-dateutil==2.4.2
python-mimeparse==0.1.4
python-obscene-words-filter==0.1.6
pytils  # pytils==0.2.3
pytz==2013b0
PyYAML==3.10
rcssmin==1.0.6
redis==2.10.5
requests==2.0.0
requests-oauthlib==0.3.2
rjsmin==1.1.0
rq==0.3.7
rq-dashboard==0.3.1
scandir==1.10.0
simplegeneric==0.8.1
simplejson  # simplejson==2.6.2
six==1.12.0
sorl-thumbnail==11.12
sqlparse==0.3.0
statsd==2.1.2
times==0.6.1
traitlets==4.3.2
twython==3.1.0
typing==3.7.4.1
ujson==1.33
urwid==1.1.1
watchdog==0.6.0
wcwidth==0.1.7
Werkzeug==0.8.3
